# PERF: setting values via df.loc / df.iloc with pyarrow-backed columns