@mroeschke 2022-12-17
Okay glad to see that Arrow is erroring in this case. Could you add a test this invalid setting case?

Yeah I think generally `ArrowInvalid: Could not convert 'foo' with type str: tried to convert to int64` is a better error message but the refactoring can be left for a separate PR