@lukemanley 2022-12-17
> Okay glad to see that Arrow is erroring in this case. Could you add a test this invalid setting case?

Added